+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Next
 
+* Add basepower: filter and is:goldborder filter.
+* Fix filtering in D1.
+* Add a button to clear the current search.
+
 # 4.21.0
 
 * Community reviews (for weapons and armor) are in for Destiny 2 inventory.